(cmake) Update all libraries to C++17. Remove Boost from the project libraries.

+++ app_support/include/vc/app_support/ProgressIndicator.hpp
@@ -184,7 +184,9 @@ public:
     {
         auto begin = std::begin(container_);
         auto end = std::end(container_);
-        return iterator{begin, std::distance(begin, end), label_, useColors_};
+        return iterator{
+            begin, static_cast<size_t>(std::distance(begin, end)), label_,
+            useColors_};
     }
 
     /**
